requirements.txt
streamlit
networkx
python-louvain
python-igraph
pandas
numpy
pyvis
plotly
scikit-learn
